third_party/blink/web_tests/virtual/controls-refresh/calendar-picker/date-picker-today-hotkey.html

<!DOCTYPE html>
<html>
<head>
<script src="../../../resources/testharness.js"></script>
<script src="../../../resources/testharnessreport.js"></script>
<script src="../../../fast/forms/resources/common.js"></script>
<script src="../../../fast/forms/resources/picker-common.js"></script>
<script src="../../../fast/forms/calendar-picker/resources/calendar-picker-common.js"></script>
</head>
<body>
<input type="date" id="date" value="2019-02-14">
<script>

promise_test(() => {
  let dateElement = document.getElementById('date');
  return openPickerWithPromise(dateElement)
  .then(() => {
    eventSender.keyDown('t');

    let splitDate = dateElement.value.split('-');
    let actualTodayDateString = new Date(splitDate[0], splitDate[1] - 1, [splitDate[2]]).toDateString();
    let expectedTodayDateString = new Date().toDateString();

    assert_equals(actualTodayDateString, expectedTodayDateString, "Today hotkey should have updated in-page control to today's date");
    assert_equals(internals.pagePopupWindow, null, "'t' key should close popup.");
  });
}, "Date picker: 't' key should select today's date and close popup");

</script>
</body>
</html>
